Update referencing.
Use the package "cleveref" for references, replacing \link by \cref.
This semi-breaks links to the bibliography: the location is still
correct, but the displayed name is always 'Section 5.5.7'. But the
bibliography needs to be redone (with biblatex) anyway, so I don’t care
too much about this.

Use the package "caption" to ensure that references to figures, tables,
etc. go to the top of the environment instead to the bottom (where the
label needs to be placed, underneath the caption).